horizontal recycler view deleted and image working

+++ app/src/main/java/com/example/myapplication/network/modules/MovieResponse.kt
@@ -21,7 +21,7 @@ data class MoviesDetails(
     @SerializedName("original_language")
     val originalLanguage: String,
     @SerializedName("vote_average")
-    val voteAverage: Int,
+    val voteAverage: Float,
     @SerializedName("overview")
     val overview: String,
     @SerializedName("release_date")
